Fix lint and typecheck errors
