+++ packages/loot-core/src/client/actions/index.ts
@@ -1,5 +1,4 @@
 export * from './queries';
-export * from './account';
 export * from './modals';
 export * from './notifications';
 export * from './prefs';
