+++ packages/loot-core/src/client/actions/queries.ts
@@ -5,6 +5,11 @@ import throttle from 'throttleit';
 import { send } from '../../platform/client/fetch';
 import { type AccountEntity } from '../../types/models';
 import * as constants from '../constants';
+import {
+  type MarkAccountReadAction,
+  type SetLastTransactionAction,
+  type UpdateNewTransactionsAction,
+} from '../state-types/queries';
 import { type AppDispatch, type GetRootState } from '../store';
 
 import { pushModal } from './modals';
@@ -366,6 +371,101 @@ export function reopenAccount(accountId) {
 
 export function forceCloseAccount(accountId) {
   return closeAccount(accountId, null, null, true);
+}
+
+// Remember the last transaction manually added to the system
+export function setLastTransaction(
+  transaction: SetLastTransactionAction['transaction'],
+): SetLastTransactionAction {
+  return {
+    type: constants.SET_LAST_TRANSACTION,
+    transaction,
+  };
+}
+
+export function parseTransactions(filepath, options) {
+  return async () => {
+    return await send('transactions-parse-file', {
+      filepath,
+      options,
+    });
+  };
+}
+
+export function importPreviewTransactions(id: string, transactions) {
+  return async (dispatch: AppDispatch): Promise<boolean> => {
+    const { errors = [], updatedPreview } = await send('transactions-import', {
+      accountId: id,
+      transactions,
+      isPreview: true,
+    });
+
+    errors.forEach(error => {
+      dispatch(
+        addNotification({
+          type: 'error',
+          message: error.message,
+        }),
+      );
+    });
+
+    return updatedPreview;
+  };
+}
+
+export function importTransactions(id: string, transactions, reconcile = true) {
+  return async (dispatch: AppDispatch): Promise<boolean> => {
+    if (!reconcile) {
+      await send('api/transactions-add', {
+        accountId: id,
+        transactions,
+      });
+
+      return true;
+    }
+
+    const {
+      errors = [],
+      added,
+      updated,
+    } = await send('transactions-import', {
+      accountId: id,
+      transactions,
+      isPreview: false,
+    });
+
+    errors.forEach(error => {
+      dispatch(
+        addNotification({
+          type: 'error',
+          message: error.message,
+        }),
+      );
+    });
+
+    dispatch({
+      type: constants.SET_NEW_TRANSACTIONS,
+      newTransactions: added,
+      matchedTransactions: updated,
+      updatedAccounts: added.length > 0 ? [id] : [],
+    });
+
+    return added.length > 0 || updated.length > 0;
+  };
+}
+
+export function updateNewTransactions(changedId): UpdateNewTransactionsAction {
+  return {
+    type: constants.UPDATE_NEW_TRANSACTIONS,
+    changedId,
+  };
+}
+
+export function markAccountRead(accountId): MarkAccountReadAction {
+  return {
+    type: constants.MARK_ACCOUNT_READ,
+    accountId,
+  };
 }
 
 const _undo = throttle(() => send('undo'), 100);
